Refactor db connection into model file
